--- src/Canino.java
@@ -1,3 +1,0 @@
-public interface Canino {
-  //TODO: Realizar la definicon de la clase
-}
