--- src/Perro.java
@@ -1,10 +1,0 @@
-public class Perro extends Mascota implements Canino{
-    RazaPerro raza;
-    public Perro(String paisOrige, double costo, String nombre, byte edad, RazaPerro raza){
-        super(paisOrige, costo, nombre, edad);
-        this.raza = raza;
-    }
-  
-    
-  
-}
