--- src/RazaPerro.java
@@ -1,3 +1,0 @@
-public enum RazaPerro {
-    //TODO:Definir razas
-}
