+++ src/models/Canino.java
@@ -1,4 +1,4 @@
-public interface Felino {
+package models;
+public interface Canino {
   //TODO: Realizar la definicon de la clase
-    
 }
