+++ src/models/Felino.java
@@ -1,0 +1,6 @@
+package models;
+
+public interface Felino {
+  //TODO: Realizar la definicon de la clase
+    
+}
