+++ src/models/Gato.java
@@ -1,4 +1,7 @@
+package models;
+
 public class Gato extends Mascota implements Felino{
+  
     RazaGato raza;
     public Gato(String paisOrige, double costo, String nombre, byte edad, RazaGato raza) {
         super(paisOrige, costo, nombre, edad);
