+++ src/models/Paises.java
@@ -1,0 +1,4 @@
+package models;
+public enum Paises {
+
+}
