+++ src/models/Perro.java
@@ -1,0 +1,9 @@
+package models;
+public class Perro  implements Canino{
+    RazaPerro raza;
+    public Perro(String paisOrige, double costo, String nombre, byte edad){
+        // super(paisOrige, costo, nombre, edad);
+       
+    }
+  
+}
