+++ src/models/RazaGato.java
@@ -1,3 +1,4 @@
+package models;
 public enum RazaGato {
     //TODO:Definir razas
 }
