+++ src/models/RazaPerro.java
@@ -1,0 +1,7 @@
+package models;
+public enum RazaPerro {
+   //TODO:Definir razas
+}
+
+
+
